pyproject.toml
[tool.poetry]
name = "alcounting-bot"
#version = "0.1.0"
description = "A python telegram bot that helps to host parties."
authors = ["nikitamorozov <nikitwww@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
python-telegram-bot = "^21.4"
toml = "^0.10.2"

[tool.poetry.group.dev.dependencies]
commitizen = "^4.1.0"
pylint = "^3.3.2"
flake8 = "^7.1.1"
black = "^24.10.0"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
pytest-flake8 = "^1.3.0"
pytest-cov = "^6.0.0"
pytest-sugar = "^1.0.0"

[tool.pytest.ini_options]
addopts = "--flake8"
flake8-max-line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#[tool.commitizen]
#version_provider = "poetry"
#name = "cz_conventional_commits"
#update_changelog_on_bump = true
#style = [
#    ["qmark", "fg:#ff9d00 bold"],
#    ["question", "bold"],
#    ["answer", "fg:#ff9d00 bold"],
#    ["pointer", "fg:#ff9d00 bold"],
#    ["highlighted", "fg:#ff9d00 bold"],
#    ["selected", "fg:#cc5454"],
#    ["separator", "fg:#cc5454"],
#    ["instruction", ""],
#    ["text", ""],
#    ["disabled", "fg:#858585 italic"]
#]

